Improve & Refactor Table Rendering
Add a `Table` module & refactor the Gains & Trade List modules to use it
to generate their tables.

Add a viewport for the table body along with an update function for
scrolling it using vim keys.

Add borders between each table row.

Calculate the widths of each column based on it's contents and a given
weight.