Add Ability to Fetch Historical GDAX & Binance Prices
Add `getHistoricalPrice` functions to the GDAX & Binance modules for
fetching the price of a currency at some point in the past.

Modify `priceUpdateChannel` to subscribe to all GDAX USD markets.

Refactor the currency out of the `GDAX.connect` function, allowing
subscriptions to different GDAX USD markets.

Refactor the Binance module to handle coins with no ETH market, although
the list of coins is currently hardcoded.