Refactor Gains Table From EthereumGains & Add FIFO Queues
Add the QuantityQueue module with the Queue datatype for tracking the
per-transaction amount & cost basis of currency purchases as well as any
gains relaized from sales. The module contains addPurchase, addSale, &
addFee functions for manipulating a queue as well as functions for
pulling out the totals. Fees are incorporated into a purchase's cost
basis instead of becoming a realized loss.

Refactor the calculations & table rendering for the Ethereum gains table
into a separate GainsTable module. This module manages a mapping from
Currencies to QuantityQueues. A sequence of Transactions is used to
update the queues & then the Currency & Aggregated data is calculated
from the final queues.