Add USD Portfolio / Gains Table
Add the USDGains module which implements a GainsTable for all
Currencies in USD prices.

On initialization, the module fetches the historical USD prices for each
Transaction and stores it in a PriceCache. This cache is stored in a
serialized file & loaded on startup.

The module uses the historical prices to build the GainsTable.Queues for
every Transaction & renders the final state as a Gains Table.

Add Binary instances for Currency & Quantity so they can be serialized
with the PriceCache.