Apply Extra Spacing to Multiple Columns
Any extra spacing when rendering a table row is equally divided among
all the columns. However, when there is space that cannot equally divide
into the number of columns, the remaining space is added to the first
column.

This changes that behavior so the extra space is distributed to as many
columns as possible, starting from the first one.